src/requirements.txt
python-dotenv
google-cloud-dlp
google-cloud-speech
awscli
boto3
ginza